updated to v0.1.4 - added basic unit tests and separated out main script into separate files. Fixed issue with sliders not working properly
